Shapes completely removed
