+++ src/main/kotlin/com/manimdsl/linearrepresentation/VariableNameGenerator.kt
@@ -2,18 +2,13 @@ package com.manimdsl.linearrepresentation
 
 import com.manimdsl.frontend.ErrorType
 import com.manimdsl.frontend.SymbolTableVisitor
-import com.manimdsl.shapes.Shape
 
 interface NameGenerator {
-    fun generateShapeName(shape: Shape): String
-
     fun generateNameFromPrefix(prefix: String): String
 }
 
 class VariableNameGenerator(private val symbolTableVisitor: SymbolTableVisitor) : NameGenerator {
     private val prefixCounter: MutableMap<String, Int> = mutableMapOf()
-
-    override fun generateShapeName(shape: Shape): String = generateNameFromPrefix(shape.pythonVariablePrefix)
 
     override fun generateNameFromPrefix(prefix: String): String {
         var count = prefixCounter.getOrDefault(prefix, 0)
@@ -26,10 +21,6 @@ class VariableNameGenerator(private val symbolTableVisitor: SymbolTableVisitor) 
 }
 
 object DummyNameGenerator : NameGenerator {
-    override fun generateShapeName(shape: Shape): String {
-        return "dummy"
-    }
-
     override fun generateNameFromPrefix(prefix: String): String {
         return "dummy"
     }
